requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
boto3==1.33.6
botocore==1.33.6
celery==5.4.0
certifi==2023.11.17
charset-normalizer==3.3.2
chime==0.7.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.7
dj-database-url==2.1.0
Django==4.2.7
django-bitcoin==0.2
django-cc==0.2.3
django-db==0.0.7
django-environ==0.11.2
django-import-export==3.3.3
django-paypal==2.0
django-resized==1.0.2
django-storages==1.14.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
filelock==3.13.1
gunicorn==21.2.0
idna==3.6
jmespath==1.0.1
kombu==5.3.7
MarkupPy==1.14
mock==5.1.0
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
Pillow==10.1.0
platformdirs==4.0.0
prompt-toolkit==3.0.43
psycopg==3.1.14
psycopg-binary==3.1.14
pycoin==0.92.20230326
PyMySQL==1.1.0
python-bitcoinrpc==1.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
reportlab==4.0.7
requests==2.31.0
s3transfer==0.8.2
setuptools==69.0.2
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.25.4
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.13
wheel==0.42.0
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0